Fix StatsigClient memory leak from repeat clients (SSR) of SessionReplay (#144)
Moving `_subscribeToVisiblityChanged()` from the `SessionReplay` constructor to a separate function. This isolates the closure scope so it no longer captures `this` in its closure (which previously happened because of a sibling lambda function which referenced `this`).

This is a short-term fix to remove the memory leak of the `StatsigClient`. Ideally we fix this further by shutting down this client post-SSR, and cleaning up all the listeners (preventing the leaky listeners which was leading to the leaky client).

## Test Plan
Tested using this update and spamming a cURL to trigger SSR a ton of times. Before this change, memory grew and contained a ton of references to SessionReplay. With this change, that no longer happens.