Fix EventLogger process hangs in Node.js
If the client SDK is used on a server environment enabling event logging causes
processes to hang indefintely.

This is due to the `setInterval` call in `EventLogger` stopping Node.js from
exiting.

This diff adds a `Timeout::unref` call in the `EventLogger` so processes can
exit normally.

Events can be flushed by adding a `beforeExit` hook and handling termination.
There might be another similar problem on the network code as trying to flush
on a beforeExit hook also causes processes to hang.