chore: remove forbidden Error mutation (#674)
A customer started getting the following error after the QuotaExceededError logic:

`Cannot set property message of  which has only a getter`

And this is the only place we set `.message` in the SDK. I tested in Chrome, and setting `error.message` doesn't update the value returned by `error.message`. So I'm removing this logic.