Event logger P5 - Event Logger (#724)
This PR contains the **event logger** and newly created **EventSender** class.
EventLogger is now purely for preprocessing, and delegating.

I created EventSender as it felt wrong to have the networking functionalities in the event logger. This draws a much more clear line for what EventLogger is in charge of. 

Regarding saving events to local storage. Currently, we enqueue all events and check during the send if logging is enabled. So, once that flag flips, we will log everything enqueued unless we overflow or retry too many times in which case we will drop them. This is all done in memory. If it is preferred to go the local storage way like the old implementation i can do that. But let's do that after everything else is reviewed to keep it simple. I kept all the storage methods in the event logger in case we do want that implementation.