fix: add functionality for saving to local storage (#729)
### TL;DR

Integrated functionality to new event logger design to store logged events in local storage when logging is disabled, allowing events to be stored locally and flushed over network when logging is re-enabled.

### How it works

- When logging is disabled, events are not enqueued to the global pending events queue. It is instead stored in local storage
- If the max amount of events are reached in local storage, we drop the oldest ones
- When logging switches to enabled, the events are loaded from memory and enqueued in the global pending events queue. Here it goes through the regular flow of getting batched and flushed


### Why make this change?

This change is important to not lose events while logging is disabled. This functionality exists in the current event logger design. However, it needed to be re-implemented as  the entire flushing mechanism and flow has changed.